+++ nodejshelper/modules/lhnodejshelper/tokenvisitor.php
@@ -12,7 +12,7 @@ if (is_numeric($chatId)) {
 
     if ($chat instanceof erLhcoreClassModelChat && $chat->hash == $Params['user_parameters']['hash'] && $chat->status !== erLhcoreClassModelChat::STATUS_CLOSED_CHAT) {
         echo json_encode([
-            'instance_id' => (erLhcoreClassModule::getExtensionInstance('erLhcoreClassExtensionNodejshelper')->getSettingVariable('automated_hosting') ? erLhcoreClassInstance::getInstance()->id : 0),
+            'instance_id' => (erLhcoreClassModule::getExtensionInstance('erLhcoreClassExtensionNodejshelper')->getSettingVariable('automated_hosting') ? (int)erLhcoreClassInstance::getInstance()->id : 0),
             'hash' => sha1($date . 'Visitor' . erConfigClassLhConfig::getInstance()->getSetting('site','secrethash') .'_' . $chatId) . '.' . $date
         ]);
     } else {
@@ -21,7 +21,7 @@ if (is_numeric($chatId)) {
 
 } else {
     echo json_encode([
-        'instance_id' => (erLhcoreClassModule::getExtensionInstance('erLhcoreClassExtensionNodejshelper')->getSettingVariable('automated_hosting') ? erLhcoreClassInstance::getInstance()->id : 0),
+        'instance_id' => (erLhcoreClassModule::getExtensionInstance('erLhcoreClassExtensionNodejshelper')->getSettingVariable('automated_hosting') ? (int)erLhcoreClassInstance::getInstance()->id : 0),
         "hash" => sha1($date . 'Visitor' . erConfigClassLhConfig::getInstance()->getSetting('site','secrethash')) . '.' . $date
     ]);
 }
